restructure packages to avoid import conflicts and receiver-type confusing, tidy up (& remove unnecessary) inline/reified
